sorting pod latency data
